feat: refactor pricing components to use PriceDisplay and improve billing details handling
