+++ core/src/zap/components/payments/billing-cards.tsx
@@ -12,10 +12,14 @@ import {
   CardTitle,
 } from "@/components/ui/card";
 import { ZapButton } from "@/components/zap-ui/button";
-import type { ProductMetadata, RecurringInterval } from "@/zap.config";
+import type { ProductMetadata } from "@/zap.config";
 import { PricingToggle } from "@/zap/components/landing/pricing/pricing-toggle";
+import { PriceDisplay } from "@/zap/components/payments/price-display";
 import { authClient } from "@/zap/lib/auth/client";
-import { getBillingDetails } from "@/zap/lib/payments/utils";
+import {
+  getProductBillingDetails,
+  getSortedProducts,
+} from "@/zap/lib/payments/utils";
 
 interface BillingCardsProps {
   products: ProductMetadata[];
@@ -43,95 +47,7 @@ export function BillingCards({ products }: BillingCardsProps) {
     }
   };
 
-  const productsArray = products.flatMap((product) => {
-    if (product.billingOptions) {
-      const key = isYearly ? "yearly" : "monthly";
-      const billing = product.billingOptions[key];
-
-      if (!billing) {
-        return [];
-      }
-
-      return {
-        ...product,
-        ...billing,
-        slug: `${product.slug}-${key}`,
-      };
-    }
-
-    if (
-      product.recurringInterval === "one-time" ||
-      !(product.billingOptions || product.recurringInterval)
-    ) {
-      return product;
-    }
-
-    return [];
-  });
-
-  const renderPrice = (price: number | string, interval: RecurringInterval) => {
-    let displayPrice: string;
-    let intervalText: string;
-
-    if (typeof price === "string") {
-      displayPrice = price;
-      intervalText = "";
-    } else if (interval === "year") {
-      displayPrice = `$${(price / 12).toFixed(2)}`;
-      intervalText = "/month";
-    } else if (interval === "month") {
-      displayPrice = `$${price.toFixed(2)}`;
-      intervalText = "/month";
-    } else if (interval === "one-time") {
-      displayPrice = `$${price.toFixed(2)}`;
-      intervalText = "one-time";
-    } else {
-      displayPrice = `$${price.toFixed(2)}`;
-      intervalText = "";
-    }
-
-    return (
-      <div className="mt-2 flex flex-col items-center space-y-1 text-center transition-all duration-500">
-        <div className="flex items-end space-x-2">
-          <span className="text-4xl font-extrabold tracking-tight">
-            {displayPrice}
-          </span>
-          {intervalText && (
-            <span className="text-muted-foreground mb-1 text-sm font-medium">
-              {intervalText}
-            </span>
-          )}
-        </div>
-      </div>
-    );
-  };
-
-  const getProductBillingDetails = (
-    productData: ProductMetadata,
-    yearlyMode: boolean,
-  ) => {
-    const billingKey = yearlyMode ? "yearly" : "monthly";
-    const billingOption = productData.billingOptions?.[billingKey];
-
-    return {
-      price: productData.price ?? billingOption?.price ?? 0,
-      recurringInterval:
-        productData.recurringInterval ??
-        billingOption?.recurringInterval ??
-        "month",
-    };
-  };
-
-  const sortedProducts = productsArray.sort((a, b) => {
-    const aPrice = getBillingDetails(a, isYearly).price;
-    const bPrice = getBillingDetails(b, isYearly).price;
-
-    if (typeof aPrice === "string" || typeof bPrice === "string") {
-      return 0;
-    }
-
-    return aPrice - bPrice;
-  });
+  const sortedProducts = getSortedProducts(products, isYearly);
 
   return (
     <div className="w-full">
@@ -161,7 +77,11 @@ export function BillingCards({ products }: BillingCardsProps) {
 
               <CardHeader className="text-center">
                 <CardTitle className="text-2xl">{product.name}</CardTitle>
-                {renderPrice(price, recurringInterval)}
+                <PriceDisplay
+                  alignment="center"
+                  interval={recurringInterval}
+                  price={price}
+                />
                 <CardDescription className="text-base">
                   {product.description}
                 </CardDescription>
